ai_conversations_summary/requirements.txt
# Core runtime dependencies (GUI + Web + RAG)
# Note: sentence-transformers will install torch and related wheels,
# so first-time installation may take significantly longer.
openai>=1.0.0
numpy
faiss-cpu
sentence-transformers
transformers>=4.37.0
tkcalendar
fastapi
uvicorn
jinja2
langchain-tavily

# Optional in this subproject, but useful in shared workspace where
# nav_dashboard exposes image-upload form endpoints.
python-multipart

# Optional vector backend
# chromadb

# Optional packaging
# pyinstaller

# Tip:
# One-click setup scripts are available at:
# - setup.bat (Windows)
# - setup.sh (Linux/macOS)
